requirements.txt
opencv-python
customtkinter
Pillow